Split Scala and Java LoadOrder APIs Closes #22
